+++ CHANGELOG.md
@@ -18,7 +18,7 @@ configuration is still supported but deprecated. See
 - fortios: support for FortiADC (@electrocret)
 - output/git: cache commit log to improve performance of oxidized-web. Fixes #3121 (@robertcheramy)
 - digest auth handles special characters in passwords by itself (no need to url encode them manually) (@einglasvollkakao)
-- changed the configuration for oxidized-web from rest: to extentions.oxidized-web (@robertcheramy)
+- changed the configuration for oxidized-web from rest: to extensions.oxidized-web (@robertcheramy)
 - netgear: add pager-handler workaround, fixes: #2394 and #3341 (@candlerb, @syn-bit)
 
 ### Fixed